Refactor file structure for each feature
