+++ src/components/meal-log/LogEntry.js
